[refactor/#68] 라우터 분리
