+++ verbly_fe/.env
@@ -1,2 +1,2 @@
 # .env (프로젝트 최상위 루트)
-VITE_API_BASE_URL=http://3.35.202.0:8080
+VITE_API_BASE_URL=https://3.35.202.0:443
