+++ verbly_fe/src/pages/Onboarding/LoginPage.tsx
@@ -1,8 +1,8 @@
-import { useNavigate } from 'react-router-dom';
-import Logo from '../../components/Logo/Logo';
-import googleIcon from '../../assets/emoji/google.svg?react';
-import kakaoIcon from '../../assets/emoji/kakao.svg?react';
-import SolidButton from '../../components/Button/SolidButton';
+import { useNavigate } from "react-router-dom";
+import Logo from "../../components/Logo/Logo";
+import googleIcon from "../../assets/emoji/google.svg?react";
+import kakaoIcon from "../../assets/emoji/kakao.svg?react";
+import SolidButton from "../../components/Button/SolidButton";
 
 const LoginPage = () => {
   // 뒤로가기 등의 기능을 위해 navigate를 준비했지만, 소셜 로그인은 href 이동을 사용합니다.
@@ -30,16 +30,14 @@ const LoginPage = () => {
       <div className="flex-1 h-full flex items-center justify-center">
         {/* 로그인 폼 내부 컨테이너 */}
         <div className="flex flex-col w-[460px] items-center gap-8">
-          
           {/* 로고 영역 */}
           <div className="w-[144px] h-[40px] flex justify-center items-center cursor-pointer">
-             {/* Logo 컴포넌트 내부에서 이미 navigate 처리가 되어 있으나, 필요시 여기서 감쌀 수도 있습니다. */}
+            {/* Logo 컴포넌트 내부에서 이미 navigate 처리가 되어 있으나, 필요시 여기서 감쌀 수도 있습니다. */}
             <Logo variant="hori" />
           </div>
 
           {/* 소셜 로그인 버튼 목록 */}
           <div className="flex flex-col gap-3 w-full">
-            
             {/* 구글 로그인 버튼 */}
             <SolidButton
               variant="assistive"
@@ -47,7 +45,7 @@ const LoginPage = () => {
               Icon={googleIcon}
               className="justify-center !text-gray-9 text-opacity-85 w-full border border-gray-200"
               onClick={handleGoogleLogin}
-              label='Continue with Google'
+              label="Continue with Google"
             />
 
             {/* 카카오 로그인 버튼 */}
@@ -57,7 +55,7 @@ const LoginPage = () => {
               Icon={kakaoIcon}
               className="justify-center !bg-[#FEE500] !text-black w-full hover:!bg-[#FDD835]" // 카카오 고유 컬러 적용
               onClick={handleKakaoLogin}
-              label='Continue with Kakao talk'
+              label="Continue with Kakao talk"
             />
           </div>
         </div>
